# src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: API
  version: 1.0.0

tags:
  - name: user
  - name: habit
paths:
  /v1/signup:
    post:
      tags:
        - user
      summary: Register a new user
      description: Create a new user
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
            example:
              email: "john.doe@example.com"
              password: "password123"
              firstName: "John"
              lastName: "Doe"
              username: "awesomeJohn6767"
      responses:
        '201':
          description: User successfully created
        '400':
          description: Bad Request - Validation error or user already exists
  /v1/login:
    post:
      tags:
        - user
      summary: Authenticate user
      description: Log in a user
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "john.doe@example.com"
              password: "password123"
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid credentials
  /v1/users/{username}:
    get:
      tags:
        - user
      summary: Get a user by username
      description: Returns a single user if found
      parameters:
        - name: username
          in: path
          required: true
          description: Username of the user to retrieve
          schema:
            type: string
            example: 'username'
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
  /v1/habits:
    get:
      tags:
        - habit
      summary: Get all habits
      description: Get all habits
      operationId: habitGetAll
      responses:
        '200':
          description: List of habits
          content:
            application/json:
              schema:
                type: object
                properties:
                  habits:
                    type: array
                    items:
                      $ref: '#/components/schemas/Habit'


    post:
      tags:
        - habit
      summary: Create a new habit
      description: Add a new habit for user
      operationId: habitCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HabitCreationRequest'
            example:
              name: "sit ups"
              description: "do 15 sit ups"
      responses:
        '201':
          description: Habit successfully created
  /v1/habits/{id}:
    delete:
      tags:
        - habit
      summary: Remove a habit
      description: Delete a habit
      operationId: habitDelete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Habit removed successfully
        '404':
          description: Habit not found

    put:
      tags:
        - habit
      summary: Update a habit
      description: Update a habit
      operationId: habitUpdate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HabitUpdateRequest'
      responses:
        '200':
          description: Habit updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Habit'
        '404':
          description: Habit not found
  /v1/habits/{id}/completions:
    post:
      tags:
        - habit
      summary: Log a habit completion
      description: Record that the habit was completed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HabitCompletionCreateRequest'
      responses:
        '201':
          description: Habit completion logged successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HabitCompletion'
        '404':
          description: Habit not found

    get:
      tags:
        - habit
      summary: Get history for a habit
      description: Get all records for a habit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  completions:
                    type: array
                    items:
                      $ref: '#/components/schemas/HabitCompletion'
        '404':
          description: Habit not found

  /v1/habits/{id}/streak:
    get:
      tags:
        - habit
      summary: Get streak for a habit
      description: Get streak history/statistics for a habit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  completions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Streak'
        '404':
          description: Habit not found


components:
  schemas:
    SignupRequest:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
        - username
      properties:
        email:
          type: string
          format: email
          description: User's email
          example: "john.doe@example.com"
        password:
          type: string
          format: password
          description: User's password
          example: "password123"
        firstName:
          type: string
          description: User's first name
          example: "John"
        lastName:
          type: string
          description: User's last name
          example: "Doe"
        username:
          type: string
          description: Username
          example: "awesomeJohn6767"

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email
          example: "john.doe@example.com"
        password:
          type: string
          format: password
          description: User's password
          example: "password123"

    HabitCreationRequest:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          description: Habit's name
          example: "sit ups"
        description:
          type: string
          description: Habit's description
          example: "do 15 sit ups"

    HabitUpdateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          example: "morning meditation"
        description:
          type: string
          maxLength: 500
          example: "15 minutes of meditation"
        frequency:
          type: string
          enum: [daily, weekly, custom]
          example: "daily"

    HabitCompletionCreateRequest:
      type: object
      properties:
        completedAt:
          type: string
          format: date-time
        isSkipped:
          type: boolean
          default: false

    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 123
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"
        username:
          type: string
          example: "awesomeJohn6767"
        email:
          type: string
          format: email
          example: "john.doe@example.com"

    Habit:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 123
        name:
          type: string
          example: "sit ups"
        description:
          type: string
          example: "do 15 sit ups"
        frequency:
          type: string
          enum: [daily, weekly, custom]
          example: "daily"
        createdAt:
          type: string
          format: date-time
          example: "23/12/2025 12:56"

    HabitCompletion:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 123
        habitId:
          type: integer
          format: int64
          example: 222
        completedAt:
          type: string
          format: date-time
          example: "22/03/2025 10:26"
        isSkipped:
          type: boolean
          default: false

    Streak:
      type: object
      properties:
        id:
          type: number
          example: 11
        currentStreak:
          type: integer
          format: int64
          example: 12
        longestStreak:
          type: integer
          format: int64
          example: 35